+++ async_upnp_client/search.py
@@ -4,24 +4,21 @@ import asyncio
 import logging
 from asyncio import DatagramTransport
 from asyncio.events import AbstractEventLoop
-from ipaddress import ip_address
+from ipaddress import IPv4Address, IPv6Address
 from typing import Awaitable, Callable, Optional, cast
 
 from async_upnp_client.const import SsdpSource
 from async_upnp_client.ssdp import (
     SSDP_DISCOVER,
-    SSDP_IP_V4,
-    SSDP_IP_V6,
     SSDP_MX,
-    SSDP_PORT,
     SSDP_ST_ALL,
     AddressTupleVXType,
     IPvXAddress,
     SsdpHeaders,
     SsdpProtocol,
     build_ssdp_search_packet,
+    determine_source_target,
     get_host_string,
-    get_source_ip_from_target_ip,
     get_ssdp_socket,
 )
 
@@ -35,7 +32,7 @@ class SsdpSearchListener:  # pylint: disable=too-many-arguments,too-many-instanc
         self,
         async_callback: Callable[[SsdpHeaders], Awaitable],
         loop: Optional[AbstractEventLoop] = None,
-        source_ip: Optional[IPvXAddress] = None,
+        source: Optional[AddressTupleVXType] = None,
         target: Optional[AddressTupleVXType] = None,
         timeout: int = SSDP_MX,
         service_type: str = SSDP_ST_ALL,
@@ -45,10 +42,9 @@ class SsdpSearchListener:  # pylint: disable=too-many-arguments,too-many-instanc
         self.async_callback = async_callback
         self.async_connect_callback = async_connect_callback
         self.service_type = service_type
-        self.source_ip = source_ip
+        self.source, self.target = determine_source_target(source, target)
         self.timeout = timeout
         self.loop = loop
-        self._target: Optional[AddressTupleVXType] = target
         self._target_host: Optional[str] = None
         self._transport: Optional[DatagramTransport] = None
 
@@ -57,12 +53,11 @@ class SsdpSearchListener:  # pylint: disable=too-many-arguments,too-many-instanc
     ) -> None:
         """Start an SSDP search."""
         assert self._target_host is not None, "Call async_start() first"
-        assert self._target is not None, "Call async_start() first"
-        packet = build_ssdp_search_packet(self._target, self.timeout, self.service_type)
+        packet = build_ssdp_search_packet(self.target, self.timeout, self.service_type)
 
         assert self._transport is not None
         protocol = cast(SsdpProtocol, self._transport.get_protocol())
-        target = override_target or self._target
+        target = override_target or self.target
         protocol.send_ssdp_packet(packet, target)
 
     async def _async_on_data(self, request_line: str, headers: SsdpHeaders) -> None:
@@ -88,31 +83,30 @@ class SsdpSearchListener:  # pylint: disable=too-many-arguments,too-many-instanc
         if self.async_connect_callback:
             await self.async_connect_callback()
 
+    @property
+    def target_ip(self) -> IPvXAddress:
+        """Get target IP."""
+        if len(self.target) == 4:
+            return IPv6Address(self.target[0])
+
+        return IPv4Address(self.target[0])
+
     async def async_start(self) -> None:
         """Start the listener."""
-        if self._target is None:
-            if self.source_ip and self.source_ip.version == 6:
-                self._target = (SSDP_IP_V6, SSDP_PORT)
-            else:
-                self._target = (SSDP_IP_V4, SSDP_PORT)
+        _LOGGER.debug("Start listening for search responses")
 
-        target_ip: IPvXAddress = ip_address(self._target[0])
-
-        if self.source_ip is None:
-            self.source_ip = get_source_ip_from_target_ip(target_ip)
         # We use the standard target in the data of the announce since
         # many implementations will ignore the request otherwise
-        sock, source, _ = get_ssdp_socket(
-            self.source_ip, target_ip, port=self._target[1]
-        )
-
-        if not target_ip.is_multicast:
-            self._target_host = get_host_string(self._target)
-        else:
-            self._target_host = ""
+        sock, source, _target = get_ssdp_socket(self.source, self.target)
 
         _LOGGER.debug("Binding to address: %s", source)
         sock.bind(source)
+
+        if not self.target_ip.is_multicast:
+            self._target_host = get_host_string(self.target)
+        else:
+            self._target_host = ""
+
         loop = self.loop or asyncio.get_event_loop()
 
         await loop.create_datagram_endpoint(
@@ -132,9 +126,9 @@ async def async_search(
     async_callback: Callable[[SsdpHeaders], Awaitable],
     timeout: int = SSDP_MX,
     service_type: str = SSDP_ST_ALL,
-    source_ip: Optional[IPvXAddress] = None,
-    loop: Optional[AbstractEventLoop] = None,
+    source: Optional[AddressTupleVXType] = None,
     target: Optional[AddressTupleVXType] = None,
+    loop: Optional[AbstractEventLoop] = None,
 ) -> None:
     """Discover devices via SSDP."""
     # pylint: disable=too-many-arguments
@@ -148,11 +142,11 @@ async def async_search(
 
     listener = SsdpSearchListener(
         async_callback,
-        loop_,
-        source_ip,
-        target,
-        timeout,
-        service_type,
+        loop=loop_,
+        source=source,
+        target=target,
+        timeout=timeout,
+        service_type=service_type,
         async_connect_callback=_async_connected,
     )