+++ CHANGES.rst
@@ -1,10 +1,12 @@
 Changes
 =======
 
-0.23.5 (unreleased)
+0.24.0 (unreleased)
 
 - Add new dummy_tv/dummy_router servers (@StevenLooman)
-- Drop python 3.6 support, add python 3.10 support
+- Drop python 3.6 support, add python 3.10 support (@StevenLooman)
+- Breaking change: Improve SSDP IPv6 support, for Python versions <3.9, due to missing IPv6Address.scope_id (@StevenLooman)
+  - `SsdpListener`, `SsdpAdvertisementListener`, `SsdpSearchListener`, `UpnpProfileDevice` now take `AddressTupleVXType` for source and target, instead of IPs
 
 
 0.23.4 (2022-01-16)